GCP: note auto storage scaling enabled
For internal MySQL we have a minimum 10GB disk and recommend a
30GB disk. For GCP however, we can stick with the default (10GB)
and simply ensure that "Enable automatic storage increases" is set.
This will let GCP Cloud SQL automatically scale storage as needed
rather than setting a higher minimum. Although this setting is
default enabled, we should note that it should be checked since
we are starting with the effective minimum amount of storage (10GB).