Fixed the CIDR for the proposed network resources
The /20 CIDR for Deployment and Services is not compatible with the CIDR mentioned in the `Configuring BOSH Director on Azure` section. Azure will give you an error :

```
The address prefix 10.0.12.0/20 in resource /subscriptions/XXXXXXXXXXXX/resourceGroups/BDC-POC/providers/Microsoft.Network/virtualNetworks/PCF/subnets/Deployment has an invalid CIDR notation. For the given prefix length, the address prefix should be 10.0.0.0/20.
```