Fixed typos and incorrect links
There was a couple typos, and the URL was not a proper link to download the Ops Manager (OM). 

Still outstanding is line 115 about how to copy an encrypted AMI, but I cannot find the correct link that describes this. It is definitely not step 9 on the URL (that has to do with syslog) that is specified.