Conflicting wording is very confusing
This section states that GCP created the service account for Ops Man and then proceeds to tell you to modify the service account that you created for Ops Man. In this circumstance it is correct to state that the user has created a service account because that is what we instruct them to do.