Updates NTP server guidance on Azure
On the Azure IAAS, the value entered in the NTP server field is ignored.
However, Ops Manager does require that there both be a value and that it
corresponds to a functioning and reachable NTP server.

So, this commit adds information about the Azure hardware (PTP) clock
and also the requirement to put a valid NTP server in the field, even
though it will be ignored later.

Signed-off-by: Maya Rosecrance <mrosecrance@pivotal.io>